+++ examples/PolarAreaChartExample.html
@@ -23,7 +23,6 @@
       var title = d3.ez.component.title().mainText("Diagram of the Causes of Mortality in the Army in The East").subText("April 1854 to March 1855");
 
       // Convert csv to d3-ez data format
-      // Rename keys
       var monthNames = [
         "January", "February", "March", "April", "May", "June",
         "July", "August", "September", "October", "November", "December"
